Drag and drop support: refactoring of image from link and image from selection https://bugs.webkit.org/show_bug.cgi?id=52496
Reviewed by Alexey Proskuryakov.

Source/WebCore: 

This work cleans up the Mac code and makes it more similar to the Windows implementation,
avoiding the use of an NSView when the FrameView can be used.
The refactoring is a necessary step towards the complete support of drag and drop
in WebKit2.
        
* page/mac/FrameMac.mm:
(WebCore::Frame::imageFromRect): Modified to use FrameView instead of NSView
to generate the image for drag.

Source/WebKit/mac: 

This work cleans up the Mac code and makes it more similar to the Windows implementation,
avoiding the use of an NSView when the FrameView can be used.
The refactoring is a necessary step towards the complete support of drag and drop
in WebKit2.

* WebCoreSupport/WebDragClient.mm:
(WebDragClient::createDragImageForLink): Added.
* WebView/WebHTMLView.mm: Removed dragImageFromLink and dragImageFromURL.
* WebView/WebHTMLViewPrivate.h: Removed dragImageFromLink and dragImageFromURL.



git-svn-id: svn://svn.chromium.org/blink/trunk@75963 bbb929c8-8fbe-4397-9dbb-9b2b20218538