Google Maps says it has horizontal scrollbars when it doesn't <rdar://problem/8978005>
Reviewed by Maciej Stachowiak.

We were updating the main frame's scrollbar state for changes to
frames other than the main frame.

* WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::contentsSizeChanged):
* WebProcess/WebCoreSupport/WebChromeClient.h:
(WebKit::WebChromeClient::WebChromeClient):
Change to bail out early if we are not looking at the main frame. Also
fix some silly typos and rename the variables to be more specific.



git-svn-id: svn://svn.chromium.org/blink/trunk@78145 bbb929c8-8fbe-4397-9dbb-9b2b20218538