<http://webkit.org/b/56304> REGRESSION(r80892): 100,000+ leaks seen on the build bot
Reviewed by Oliver Hunt.

Source/JavaScriptCore:

* API/JSClassRef.cpp:
(OpaqueJSClass::OpaqueJSClass): Don't leak any existing entry for the given name if
the class definition contains duplicates. This also removes what look to be leaks
of the StringImpl instances that are used as keys: the HashMap key type is a RefPtr
which retains / releases the instances at the appropriate time, so explicitly calling
ref is not necessary.

Tools:

* DumpRenderTree/LayoutTestController.cpp:
(LayoutTestController::staticFunctions): Remove duplicate entries.

git-svn-id: svn://svn.chromium.org/blink/trunk@81051 bbb929c8-8fbe-4397-9dbb-9b2b20218538