WKErrors need the PeerCertificateChain for certificate errors <rdar://problem/8951784> https://bugs.webkit.org/show_bug.cgi?id=56592
Reviewed by Anders Carlsson.

* Shared/WebCoreArgumentCoders.h:
Special case encoding/decoding of ResourceErrors for the mac.

* Shared/mac/PlatformCertificateInfo.h:
* Shared/mac/PlatformCertificateInfo.mm:
(WebKit::PlatformCertificateInfo::PlatformCertificateInfo):
Add constructor that take the PeerCertificateChain as CFArrayRef.

* Shared/mac/WebCoreArgumentCodersMac.mm:
(CoreIPC::encodeResourceError):
(CoreIPC::decodeResourceError):
Add encode/decode for ResourceError using the underlying NSError
and encoding/decoding as much of the userInfo as we know how to.
Right now this includes all string values and the PeerCertificateChain.



git-svn-id: svn://svn.chromium.org/blink/trunk@81387 bbb929c8-8fbe-4397-9dbb-9b2b20218538