<rdar://problem/9005982> Flash of white when a WKView with composited content moves on-screen https://bugs.webkit.org/show_bug.cgi?id=57522
Reviewed by Maciej Stachowiak.

* UIProcess/DrawingAreaProxyImpl.cpp:
(WebKit::DrawingAreaProxyImpl::exitAcceleratedCompositingMode): Invalidate the current
backing store state, so that we get an update as soon as the page enters compositing mode
next.
* WebProcess/WebPage/DrawingAreaImpl.cpp:
(WebKit::DrawingAreaImpl::sendDidUpdateBackingStoreState): Added a call to
LayerTreeHost::forceRepaint() in order to flush and synchronize the layers
before sending the update message to the UI process.



git-svn-id: svn://svn.chromium.org/blink/trunk@82543 bbb929c8-8fbe-4397-9dbb-9b2b20218538