Reviewed by Oliver Hunt.
Canvas: Use fast, approximate dirty rects for stroke()
https://bugs.webkit.org/show_bug.cgi?id=59764

No new tests, this is an optimization.

* html/canvas/CanvasRenderingContext2D.cpp:
(WebCore::CanvasRenderingContext2D::stroke): Instead of using
Path::strokeBoundingRect() to calculate the exact bounding rect
for a path stroke, get the Path::boundingRect() and inflate it by
miterLimit + lineWidth to get a slightly oversized dirty rect
in a fraction of the time.


git-svn-id: svn://svn.chromium.org/blink/trunk@90758 bbb929c8-8fbe-4397-9dbb-9b2b20218538