[Chromium] SVG text is not rendered sometimes with geometricPrecision https://bugs.webkit.org/show_bug.cgi?id=64341
Source/WebCore:

Changing the type of x and y offsets of ComplexTextControllerLinux to make it possible to treat negative offsets.

Patch by Kenichi Ishibashi <bashi@chromium.org> on 2011-07-12
Reviewed by Tony Chang.

Test: platform/chromium-linux/svg/text/text-with-geometric-precision.svg

* platform/graphics/chromium/ComplexTextControllerLinux.cpp: Changed the type of m_offsetX and m_startingY from unsigned to int.
(WebCore::ComplexTextController::ComplexTextController): Changed the type of the argument.
(WebCore::ComplexTextController::reset): Ditto.
* platform/graphics/chromium/ComplexTextControllerLinux.h:

LayoutTests:

Add a test for the regression that the text is not rendered with geometricPrecision.

Patch by Kenichi Ishibashi <bashi@chromium.org> on 2011-07-12
Reviewed by Tony Chang.

* platform/chromium-linux/svg/text/text-with-geometric-precision-expected.png: Added.
* platform/chromium-linux/svg/text/text-with-geometric-precision-expected.txt: Added.
* platform/chromium-linux/svg/text/text-with-geometric-precision.svg: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@90868 bbb929c8-8fbe-4397-9dbb-9b2b20218538