Make using lowercase parameter names for AppleConnect be a plug-in quirk https://bugs.webkit.org/show_bug.cgi?id=64638
Reviewed by Sam Weinig.

* Shared/Plugins/Netscape/mac/NetscapePluginModuleMac.mm:
(WebKit::NetscapePluginModule::determineQuirks):
Set the WantsLowercaseParameterNames quirk for the AppleConnect plug-in.

* Shared/Plugins/PluginQuirks.h:
Add WantsLowercaseParameterNames quirk.

* WebProcess/Plugins/Netscape/NetscapePlugin.cpp:
(WebKit::NetscapePlugin::initialize):
If the plug-in has the WantsLowercaseParameterNames quirk, convert the parameter
names to lowercase.

* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::createPlugin):
Remove the code that would convert the parameters here. Also remove the FIXME; plug-in quirks
aren't really the same thing as site-specific quirks.


git-svn-id: svn://svn.chromium.org/blink/trunk@91188 bbb929c8-8fbe-4397-9dbb-9b2b20218538