[GTK] Change webview API tests to use "load-status" signal instead of "load-progress" https://bugs.webkit.org/show_bug.cgi?id=66243
Reviewed by Martin Robinson.

The idle_quit_loop_cb() already checks the "load-status" flag, so it is enough to
call it only when the "load-status" change.

This change is important in the context of bug 28851, which aims to change the
order in which "progress finished" and "load finished" callbacks are called. In
this new scenario, at the progress callback the load will never be finished.

* tests/testwebview.c:
(test_webkit_web_view_grab_focus):

git-svn-id: svn://svn.chromium.org/blink/trunk@93311 bbb929c8-8fbe-4397-9dbb-9b2b20218538