[Qt] Button text is invisible when mobile theme is in use and application palette set to use light text color. https://bugs.webkit.org/show_bug.cgi?id=66635
When mobile theme is in use and application palette is using light
color to display text button text becomes invisible. The reason for
that is that in RenderThemeQt::systemFont() we are using application
pallete while when are painting buttons lightGrayPalette is used.
Calling setPaletteFromPageClientIfExists() takes care of the issue.

Patch by Misha Tyutyunik <michael.tyutyunik@nokia.com> on 2011-08-23
Reviewed by Andreas Kling.

Covered by existing tests.

* platform/qt/RenderThemeQt.cpp:
(WebCore::RenderThemeQt::systemColor):

git-svn-id: svn://svn.chromium.org/blink/trunk@93652 bbb929c8-8fbe-4397-9dbb-9b2b20218538