webkit-patch rebaseline-expectations don't work when run outside the root https://bugs.webkit.org/show_bug.cgi?id=66966
Reviewed by Eric Seidel.

Turns out tool.path() used to be relative, which caused problems when
running webkit-patch outside the root directory.  This patch makes it
absolute.  I checked all callers, and they all appear to want this
behavior.

It was unclear to me how to test this change because we only really
have a testing strategy for webkitpy, not for the main script.

* Scripts/webkit-patch:


git-svn-id: svn://svn.chromium.org/blink/trunk@93829 bbb929c8-8fbe-4397-9dbb-9b2b20218538