[EFL] Do not allocate memory for extremely large surfaces. https://bugs.webkit.org/show_bug.cgi?id=65192
Patch by Raphael Kubo da Costa <kubo@profusion.mobi> on 2011-09-06
Reviewed by Martin Robinson.

So far, RenderThemeEfl tried to allocate a buffer and a cairo surface
the size of the whole form element passed to it.

In the case of
fast/overflow/overflow-height-float-not-removed-crash.html and others,
this meant extremely large widgets, which crashed the code.

We now only render the widgets if they are smaller than some hardcoded
and sufficiently large values which should work in most cases.

No new tests, as this was uncovered by existing ones.

* platform/efl/RenderThemeEfl.cpp:
(WebCore::RenderThemeEfl::isFormElementTooLargeToDisplay):
(WebCore::RenderThemeEfl::cacheThemePartNew):
(WebCore::RenderThemeEfl::paintThemePart):
* platform/efl/RenderThemeEfl.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@94578 bbb929c8-8fbe-4397-9dbb-9b2b20218538