[Chromium] Crash in WebCore::DatabaseObserver https://bugs.webkit.org/show_bug.cgi?id=67805
Patch by Stephen Chenney <schenney@chromium.org> on 2011-09-23
Reviewed by David Levin.

* src/DatabaseObserver.cpp:
(WebCore::DatabaseObserver::canEstablishDatabase): Added a check for a
null frame or page, and return false if null. Investigated
changing the fall-through return value to false but decided against
it given the way the code is used and existing default values for
related code. Reproduction and testing depends on having a document
with no frame or no page, which sometimes happens in practice but is
hard to construct explicitly.

git-svn-id: svn://svn.chromium.org/blink/trunk@95871 bbb929c8-8fbe-4397-9dbb-9b2b20218538