[Qt][WK2] Touch mocking is broken with Qt 5 post refactor merge https://bugs.webkit.org/show_bug.cgi?id=69617
Reviewed by Andreas Kling.

Injecting fake touch events using qt_translateRawTouchEvent does not work anymore, as it sends
the touch events to widgets only. The QML view however is a QWindow. Therefore the fake touch
events have to be dispatched through QWindowSystemInterface::handleTouchEvent. As a result the
events appear to be spontaenous, which required adjusting the are-we-on-a-real-touch-device
logic.

* MiniBrowser/qt/MiniBrowserApplication.cpp:
(MiniBrowserApplication::MiniBrowserApplication):
(MiniBrowserApplication::notify):
* MiniBrowser/qt/MiniBrowserApplication.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@96931 bbb929c8-8fbe-4397-9dbb-9b2b20218538