Tiled Backing Store does not regenerate tiles when it should https://bugs.webkit.org/show_bug.cgi?id=57798
Reviewed by Darin Adler.

Source/WebCore:

Changed TiledBackingStore::adjustVisibleRect to take into account
contentsSize, the same way as it is done when tiles are generated.

This is an issue that requires a certain sequence of API calls which
may not be easily simulated from DumpRenderTree, but which is
easily reproduced with QtTest. This is why I have provided only
a Qt specific test despite this being a generic problem.

* platform/graphics/TiledBackingStore.cpp:
(WebCore::TiledBackingStore::adjustVisibleRect):
Changed to take into account ContentsSize.
(WebCore::TiledBackingStore::visibleContentsRect):
Added to return the intersection of the viewport's visible rect with
the ContentsRect.
(WebCore::TiledBackingStore::createTiles):
Changed to take into account the ContentsSize when calculating the
previously visible rect.
* platform/graphics/TiledBackingStore.h:

Source/WebKit/qt:

Provided test for this bug. Changed the name of the resource I added
for a previous test so that the name is descriptive such that it can
be shared across several tests.

* tests/qgraphicswebview/tst_qgraphicswebview.cpp:
(tst_QGraphicsWebView::bug57798):
(tst_QGraphicsWebView::bug56929):
* tests/qgraphicswebview/tst_qgraphicswebview.qrc:


git-svn-id: svn://svn.chromium.org/blink/trunk@97945 bbb929c8-8fbe-4397-9dbb-9b2b20218538