[Qt] Change semantics of the haveQt() function to match API promises
Having Qt 5 does not imply that we have Qt 4, from an API point of
view, as a major version can in theory break BC/SC. Our minimum
version of Qt 4 will always be the latest released version of Qt 4,
so the only place we should use haveQt(4, x) is when checking for
the minimum requirements.

Reviewed-by Simon Hausmann.

git-svn-id: svn://svn.chromium.org/blink/trunk@99843 bbb929c8-8fbe-4397-9dbb-9b2b20218538