REGRESSION: rtl horizontal scrollbar / resize bug - Body shifts on resize when scrolled all the way to the left https://bugs.webkit.org/show_bug.cgi?id=70395
Reviewed by Tony Chang.

Source/WebCore: 

This patch fixes the problem in Mac and Chromium Mac.
   
The existing test fast/dom/rtl-scroll-to-leftmost-and-resize.html seems does not really work in Mac DRT
(the browswer window is not resized). It works in Chromium-Mac. And Chromium-Mac's code
is forked from Mac.

* platform/chromium/ScrollAnimatorChromiumMac.mm:
(WebCore::ScrollAnimatorChromiumMac::immediateScrollToPoint):
* platform/mac/ScrollAnimatorMac.mm:
(WebCore::ScrollAnimatorMac::immediateScrollToPoint):

LayoutTests: 

The test still marks as FAIL in test_expectations because cg-mac might need rebase.

* platform/chromium-mac/fast/dom/rtl-scroll-to-leftmost-and-resize-expected.png: Added.
* platform/chromium-mac/fast/dom/rtl-scroll-to-leftmost-and-resize-expected.txt: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@100819 bbb929c8-8fbe-4397-9dbb-9b2b20218538